requirements.txt
selene==2.0.0rc9
pytest
allure-pytest>=2.13.3
python-dotenv>=1.0.1
requests
jsonschema
appium-python-client
pydantic-settings
python-dotenv